Solution day 16.
first part: minimum and maximum values have been hardcoded
second part: can be made a bit better, because right now we still have to manually puzzle.. (but to just get the answer, this was faster)

the indexes of all the departure fields are : [0, 2, 3, 7, 9, 18]

code will return a list with all possible index numbers per field. There will be 1 or more fields that have only 1 index for which all entries are valid. This means that we can link the field to a certain index. now we can remove that index number from all the other possible fields. do this until all fields have been figured out. (or at least all departure fields)